fix(prune): Correct number of repacks (#167)
The reference for repacking that is both used
for `--max-unused` and `--max-repack` was incomplete. In fact the packs
which are at this point are only candidates for repacking (but not yet
decided) were not taken into account. Sometimes this can be quite a lot
and therefore explains why in a max-repack=10% example less than 10% was
repacked. And in the max-unused=5% example it took the 5% of a much
smaller amount of used data and hence tried to repack too much as the
limit of still tolerated unused data is too low. All users were
affected, and this has now been fixed.

closes https://github.com/rustic-rs/rustic/issues/1062