fix: Set correct content for symlink with parent snapshot (#174)
When having a symlink which is already present in a parent snapshot,
rustic set `content = []` instead of using the parent content. While
this does no harm (for symlinks the content is anyway not read), it may
lead to changed trees even though no change was actually performed. This
PR fixes that.