feat!: Allow specifying many options in config profile without array (#211)
Adds the possibility to specify many options in config profile files as
single strings which required arrays of strings - in the case of the
array only containing 1 entry.

Breaking change:
Also removes the splitting of one string from `password-command` and
`warmup-command`. Now the already-split command has to be given as array
(or single string if there are no arguments to the command).