fix: backup file even if failed listing extended attributes
This should fix rustic-rs/rustic_core#232

The idea:
Even on a system that supports extended-attributes,
we might encounter some part of the file-system that doesn't.

For instance a `mount` with the right option
(streams_interface=windows for instace) will not support xattr.

AFAIK it also seems to be the case with a simple sshfs on a debian/ubuntu.