fix: fix StringList::contains_all (#246)
There was a bug `SrtingList::contains_all`.
The implications of this bugs are when filtering snapshots by tags, i.e.
in rustic the `--filter-tags` option and the `--keep-tags` option of
`forget` is affected. In the latter case, too many snapshots are kept
which means this bug is non-critical.

This PR fixes this bug.