[EXPERIMENT] begin adding "async" versions of trait/impls
This is an experimentation branch.
The way I duplicated code to """quickly""" have an async impl
is a terrible way to do it.
Full code duplication is never a good option.

My goal is to reach a state of rustic/rustic_core/rustic_backend
to experiment on the async implementations.