feat: Add CommandInput (#252)
Add the new type CommandInput. This type supports serde and clap (using
`FromStr`) and is supposed to be used where users need to specify
commands which are to be executed.

There are 3 ways to define commands (here TOML is used):
- as one string: `command = "echo test"`
- as command and args as string array: `command = {command = "echo",
args = ["test1", "test2"] }`
- additionally specify error treatment: `command = {command = "echo",
args = ["test1", "test2"], on_failure = "ignore" }`

In the first example the full command and the args are split using
`shell_words` which requires special escaping or quoting when spaces are
contained. In the other examples every string is taken literally as
given (but TOML escapes may apply when parsing TOML).

This PR only adds the new type, using it on existing commands will be a
breaking change.