feat: Add Repository method to update snapshot collections (#260)
This PR adds the `update_snapshots`, `update_all_snapshots` and
`update_matching_snapshots` methods to `Repository`.

They yield the identical result as the corresponding `get_*` methods,
but allow to give a list of existing `SnapshotFile`s and won't read
those again. In interactive UIs (like rustic interactive mode) this
allows for very snapshot refreshing - which is almost costless if no or
few snapshots have changed in the repository.