fix(backend)!: Use correct merge stratgy for repository options (#291)
This fixes the merge strategy for values of `options`, `options_hot`,
and `options_cold` to ignore incoming elements if they are already
existing in the HashMap. The prior implementation was overwriting the
values.

Breaking change: The freestanding and public `extend` function has been
removed. You can use `conflate::hashmap::overwrite` as a replacement.

---------

Signed-off-by: simonsan <14062932+simonsan@users.noreply.github.com>
Co-authored-by: Alexander Weiss <alex@weissfam.de>
Co-authored-by: aawsome <37850842+aawsome@users.noreply.github.com>