feat(commands): Add convenient names for read-data-subset n/m (#328)
This PR adds convenience names for the `-read-data-subset=n/m` option
which allow for more easy regular check runs which cover the whole
repository data over a given period.

E.g. it allows to daily run `check --read-data
--read-data-subset=daily/month` which covers the repository once run at
every day of any month. Before, this requirement needed scripting to set
`n` to the number of the current day within the month and `m` to the
total count of days in the current month.

For `n`, it is now allowed to use
- `hourly`
- `daily`
- `weekly`
- `monthly`
- `yearly`
which is replaced by an index corresponding to the date/time the `check`
command is called.

For `m`, it is now allowed to use
- `day`
- `week`
- `month`
- `year`
which is depending on `n` replaced by the total count of
hours/days/weeks/months within the current entity. Moreover
`month_hours`, `month_days`, `year_hours`, `year_days` which work
independent from `n`.

Co-authored-by: simonsan <14062932+simonsan@users.noreply.github.com>