fix: improve handling of u32 conversions (#412)
As packs are only allowed up to a size of 4GiB, some internal data
structures use `u32` instead of `usize` or `u64` to save space. This PR
improves the handling of errors in the needed type conversations. As we
assume at least a 32bit systems, conversions from `u32` to `usize` will
panic.